requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-sdk==0.1.70
dataclasses-json==0.6.7
python-dateutil==2.9.0.post0
pydantic==2.11.7
faiss-cpu==1.11.0
pypdf==5.6.1
python-dotenv==1.0.1
typing_extensions==4.14.0
typing-inspect==0.9.0
